add conditional actions
a 'conditional action' performs a certain action on a layergroup based on data and a conditional expression. the conditional action is controlled by the name of the layergroup.

a common use case is to hide a layer group is a certain value in the json is true.

- the name `hide #if({isVIP})` hides the layer group is the value of the boolean data field `isVIP` is true

delete, hide, show, lock, select

`ACTION #if(CONDITION)`

the CONDITION is any javascript statement
that evaluates. the javascript statement should return a boolean value. you can use curly bracket syntax `{datafield}` to include data values in the condition.

the default action is 'show'

- `#if({isVIP})` shows the layergroup, assuming isVIP is a boolean value
- `delete #if("{name}" == "John")` name is a string
- `hide #if("{name}".length > 100)` name is a string
- `show #if({age} < 27)` age is an integer
- `select #if({isVIP})`  isVIP is a boolean value